Registrieren klappt jetzt

+++ src/main/java/ch/zli/m223/punchclock/domain/UserRole.java
@@ -9,19 +9,21 @@ import java.util.Set;
 public class UserRole {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
+
     private String name;
 
     @OneToMany(mappedBy = "userRole")
     @JsonIgnoreProperties({"userRole"})
     private Set<User> users;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
+
     }
 
     public String getName() {
